Seperate classes UndoAction and RedoAction
Issue: #20
